* Add success messages to all the forms. * Minor cleanups here and there.

+++ django_kala/projects/views/projects/settings/categories.py
@@ -1,3 +1,4 @@
+from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.core.exceptions import PermissionDenied
 from django.http import HttpResponse
@@ -16,6 +17,7 @@ class CategoriesView(LoginRequiredMixin, TemplateView):
         return {
             'categories': self.project.category_set.all(),
             'project': self.project,
+            'organization': self.project.organization,
             'can_create': self.project.has_change(self.request.user) or self.project.has_create(self.request.user),
             'can_invite': self.project.organization.has_change(self.request.user) or self.project.organization.has_create(self.request.user)
         }
@@ -27,9 +29,13 @@ class CategoriesView(LoginRequiredMixin, TemplateView):
         return super(CategoriesView, self).dispatch(request, *args, **kwargs)
 
     def delete(self, *args, **kwargs):
-        # TODO: check permissions
+        if not self.project.has_delete(self.request.user):
+            raise PermissionDenied(
+                'You do not have permission to delete this category.'
+            )
         if self.category:
             self.category.delete()
+            messages.success(self.request, 'The category has been deleted.')
             return HttpResponse(status=202)
         return self.render_to_response(self.get_context_data())
 
@@ -41,6 +47,7 @@ class NewCategoryView(LoginRequiredMixin, TemplateView):
         return {
             'form': self.form,
             'project': self.project,
+            'organization': self.project.organization,
             'can_create': self.project.has_change(self.request.user) or self.project.has_create(self.request.user),
             'can_invite': self.project.organization.has_change(self.request.user) or self.project.organization.has_create(self.request.user)
         }
@@ -56,6 +63,7 @@ class NewCategoryView(LoginRequiredMixin, TemplateView):
     def post(self, request, *args, **kwargs):
         if self.form.is_valid():
             self.form.save()
+            messages.success(request, 'The category has been saved.')
             return redirect(reverse('projects:categories', args=[self.project.pk]))
         return self.render_to_response(self.get_context_data())
 
@@ -67,6 +75,7 @@ class EditCategoryView(LoginRequiredMixin, TemplateView):
         return {
             'form': self.form,
             'project': self.project,
+            'organization': self.project.organization,
         }
 
     def dispatch(self, request, project_pk, category_pk, *args, **kwargs):
@@ -78,5 +87,6 @@ class EditCategoryView(LoginRequiredMixin, TemplateView):
     def post(self, request, *args, **kwargs):
         if self.form.is_valid():
             self.form.save()
+            messages.success(request, 'The category has been updated.')
             return redirect(reverse('projects:edit_category', args=[self.project.pk, self.category.pk]))
         return self.render_to_response(self.get_context_data())
